atom separated
