+++ lib/src/metta/runner/stdlib/random.rs
@@ -5,10 +5,7 @@ use crate::metta::*;
 use crate::space::DynSpace;
 use crate::common::shared::Shared;
 use crate::metta::text::Tokenizer;
-use crate::metta::runner::arithmetics::*;
-use crate::metta::runner::stdlib::grounded_op;
-use crate::metta::runner::Metta;
-use crate::metta::runner::stdlib::regex;
+use crate::metta::runner::{arithmetics::*, stdlib::grounded_op, Metta, stdlib::regex};
 
 use std::fmt::Display;
 use rand::Rng;
